if docker is here, let's use it

+++ packages/addons/_tests/drizzle/test.ts
@@ -48,21 +48,15 @@ beforeAll(() => {
 		dockerInstalled = false;
 	}
 
-	if (dockerInstalled && MUST_HAVE_DOCKER) {
-		execSync('docker compose up --detach', { cwd, stdio: 'pipe' });
-	}
+	if (dockerInstalled) execSync('docker compose up --detach', { cwd, stdio: 'pipe' });
 
 	// cleans up the containers on interrupts (ctrl+c)
 	process.addListener('SIGINT', () => {
-		if (dockerInstalled && MUST_HAVE_DOCKER) {
-			execSync('docker compose down --volumes', { cwd, stdio: 'pipe' });
-		}
+		if (dockerInstalled) execSync('docker compose down --volumes', { cwd, stdio: 'pipe' });
 	});
 
 	return () => {
-		if (dockerInstalled && MUST_HAVE_DOCKER) {
-			execSync('docker compose down --volumes', { cwd, stdio: 'pipe' });
-		}
+		if (dockerInstalled) execSync('docker compose down --volumes', { cwd, stdio: 'pipe' });
 	};
 });
